+++ crates/devtools/Cargo.toml
@@ -33,7 +33,7 @@ tracing-subscriber.workspace = true
 tokio.workspace = true
 tonic.workspace = true
 serde = "1"
-tonic-health = "0.10"
+tonic-health = "0.14"
 colored = "2.1"
 serde_json = "1"
 futures = "0.3.30"
