+++ crates/v1/crates/devtools/Cargo.toml
@@ -15,9 +15,9 @@ devtools-core = { path = "../../../devtools-core", version = "0.3.3" }
 tauri = { version = "1.6.1", features = ["tracing"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-tonic = "0.10"
+tonic = "0.14"
 tokio = "1.36.0"
-tonic-health = "0.10"
+tonic-health = "0.14"
 colored = "2.1"
 serde_json = "1"
 futures = "0.3.30"
