+++ crates/wire-build/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "devtools-wire-format-build"
+version = "0.1.0"
+license.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+publish = false
+
+[dependencies]
+tonic-prost-build = { version = "0.14" }
