+++ crates/wire/Cargo.toml
@@ -13,10 +13,11 @@ repository.workspace = true
 [dependencies]
 tonic.workspace = true
 prost.workspace = true
+tonic-prost.workspace = true
 prost-types.workspace = true
 tracing-core.workspace = true
 bitflags = "2.6"
 
 [dev-dependencies]
-tonic-build = "0.10"
-prost-build = "0.12"
+tonic-build = "0.14"
+prost-build = "0.14"
